Switch Example Plugin to gradle
