+++ Example-Plugin/settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name = 'Example-Plugin'
+include ':Ikala-Core'
+project(':Ikala-Core').projectDir = new File('../../Ikala-Core')
